requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Reddit API
praw==7.8.1

# Web scraping
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
feedparser==6.0.11
selenium==4.27.1  # for JS-heavy pages (optional)

# Data analysis & ML
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.6.0
scipy==1.15.0

# Scheduling
apscheduler==3.11.0

# Environment & config
python-dotenv==1.0.1
pyyaml==6.0.2

# Utilities
httpx==0.28.1
tenacity==9.0.0
rich==13.9.4

# NLP (lightweight)
nltk==3.9.1
